change DB field type to u64 remove commented code remove duplicate structs and fix return type in fn

+++ sql/insert_dummy_data.sql
@@ -42,7 +42,7 @@ CREATE TABLE IF NOT EXISTS source2target
 CREATE TABLE IF NOT EXISTS source_stats
 (
     id        serial PRIMARY KEY       NOT NULL,
-    hits      INT                      NOT NULL,
+    hits      BIGINT                      NOT NULL,
     source_id INT                      NOT NULL,
     start     TIMESTAMP WITH TIME ZONE NOT NULL,
     stop      TIMESTAMP WITH TIME ZONE NOT NULL,
@@ -53,10 +53,10 @@ CREATE TABLE IF NOT EXISTS source_stats
 CREATE TABLE IF NOT EXISTS target_stats
 (
     id        serial PRIMARY KEY       NOT NULL,
-    hits      INT                      NOT NULL,
-    min_ns    INT                      NOT NULL,
-    max_ns    INT                      NOT NULL,
-    avg_ns    INT                      NOT NULL,
+    hits      BIGINT                     NOT NULL,
+    min_ns    BIGINT                     NOT NULL,
+    max_ns    BIGINT                     NOT NULL,
+    avg_ns    BIGINT                     NOT NULL,
     start     TIMESTAMP WITH TIME ZONE NOT NULL,
     stop      TIMESTAMP WITH TIME ZONE NOT NULL,
     target_id INT                      NOT NULL,
